refactor: reorganize components and update image remote patterns
