+++ next.config.ts
@@ -3,7 +3,7 @@ import type { NextConfig } from 'next'
 const nextConfig: NextConfig = {
   /* config options here */
   images: {
-    remotePatterns: [{ hostname: 'cdn.prod.website-files.com' }],
+    remotePatterns: [{ hostname: 'cdn.prod.website-files.com' }, { hostname: 'images.unsplash.com' }],
   },
 }
