+++ src/components/layout/home/hero.tsx
@@ -1,8 +1,8 @@
-import { motion } from "framer-motion";
-import { MoveUpRight } from "lucide-react";
-import { Cover } from "../ui/cover";
-import { Highlight } from "../ui/hero-highlight";
-import { HoverBorderGradient } from "../ui/hover-border-gradient";
+import { Cover } from '@/components/ui/cover'
+import { Highlight } from '@/components/ui/hero-highlight'
+import { HoverBorderGradient } from '@/components/ui/hover-border-gradient'
+import { motion } from 'framer-motion'
+import { MoveUpRight } from 'lucide-react'
 
 export default function Hero() {
   return (
@@ -15,20 +15,8 @@ export default function Hero() {
           </span>
           <div className="relative flex space-x-2 items-center z-10 rounded-full bg-secondary/75 py-0.5 px-4 ring-1 ring-ring/10 ">
             <span>MauvaTech</span>
-            <svg
-              fill="none"
-              height="16"
-              viewBox="0 0 24 24"
-              width="16"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path
-                d="M10.75 8.75L14.25 12L10.75 15.25"
-                stroke="currentColor"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth="1.5"
-              />
+            <svg fill="none" height="16" viewBox="0 0 24 24" width="16" xmlns="http://www.w3.org/2000/svg">
+              <path d="M10.75 8.75L14.25 12L10.75 15.25" stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth="1.5" />
             </svg>
           </div>
           <span className="absolute -bottom-0 left-[1.125rem] h-px w-[calc(100%-2.25rem)] bg-gradient-to-r from-emerald-400/0 via-emerald-400/90 to-emerald-400/0 transition-opacity duration-500 group-hover:opacity-40" />
@@ -46,18 +34,11 @@ export default function Hero() {
             duration: 0.5,
             ease: [0.4, 0.0, 0.2, 1],
           }}
-          className="text-2xl px-4 md:text-4xl lg:text-5xl font-bold text-foreground  max-w-4xl leading-relaxed lg:leading-snug text-center mx-auto mt-4"
-        >
+          className="text-2xl px-4 md:text-4xl lg:text-5xl font-bold text-foreground  max-w-4xl leading-relaxed lg:leading-snug text-center mx-auto mt-4">
           Where <Highlight>Vision</Highlight> Meets <Cover>Innovation</Cover>
         </motion.h1>
-        <p className="mt-4 max-w-xl mx-auto text-sm md:text-lg text-foreground/60 text-center">
-          We build cutting-edge websites and mobile apps that drive results.
-        </p>
-        <HoverBorderGradient
-          containerClassName="rounded-full mx-auto mt-4"
-          as="button"
-          className="bg-background text-foreground flex items-center gap-2 group/button  "
-        >
+        <p className="mt-4 max-w-xl mx-auto text-sm md:text-lg text-foreground/60 text-center">We build cutting-edge websites and mobile apps that drive results.</p>
+        <HoverBorderGradient containerClassName="rounded-full mx-auto mt-4" as="button" className="bg-background text-foreground flex items-center gap-2 group/button  ">
           <span>Explore Our Work</span>
           <span>
             <MoveUpRight className="w-4 group-hover/button:rotate-45 group-hover/button:size-5 transition-all duration-300" />
@@ -65,5 +46,5 @@ export default function Hero() {
         </HoverBorderGradient>
       </div>
     </section>
-  );
+  )
 }
